[MonorepoBuilder] Abstract AbstractMergeTest
